reorganize aillio code into two separate files, and select the version to import in comm.py based on the user configuration
